refactor: rename HeroSection to Hero & move to sub dir
