+++ src/app/page.tsx
@@ -1,11 +1,11 @@
-import HeroSection from "./components/HeroSection";
+import Hero from "./components/Hero/Hero";
 import Navbar from "@/components/Navbar";
 
 export default function Home() {
   return (
     <main>
       <Navbar />
-      <HeroSection />
+      <Hero />
     </main>
   );
 }
